Moved config to own module
